+++ doc/changelog.rst
@@ -64,6 +64,8 @@ Localization
 Admin
 ~~~~~
 
+- Remove support for select pickers
+- Enable multiselect dropdown with Select2
 - Enable adding log entries to orders
 - Enable order editing
 - Allow shipment creation form extensions